backend/requirements.txt
langchain-openai==1.1.7
langchain-community==0.4.1
pypdf==6.6.0
python-dotenv==1.2.1
psycopg==3.3.2
langchain-groq==1.1.1
langchain-text-splitters==1.1.0
langchain-huggingface==1.2.0
fastapi==0.128.0
uvicorn==0.40.0
langchain-classic==1.0.1
python-multipart==0.0.21
jinja2==3.1.6
sentence-transformers==5.2.0
sqlalchemy==2.0.45
asyncpg==0.31.0
alembic==1.18.1
langchain-postgres==0.0.16
pyjwt==2.10.1
pwdlib[argon2]==0.3.0
loguru==0.7.0
pre-commit==4.5.1
